Move activity functionality into a separate module
