+++ src/adhocracy_core/adhocracy_core/__init__.py
@@ -84,10 +84,10 @@ def add_after_commit_hooks(request):
 
 def add_request_callbacks(request):
     """Add request callbacks."""
-    from adhocracy_core.auditing import update_auditlog_callback
+    from adhocracy_core.activity import update_activity_callback
     from adhocracy_core.changelog import clear_changelog_callback
     from adhocracy_core.changelog import clear_modification_date_callback
-    request.add_response_callback(update_auditlog_callback)
+    request.add_response_callback(update_activity_callback)
     request.add_finished_callback(clear_changelog_callback)
     request.add_finished_callback(clear_modification_date_callback)
